% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_chile.R
\name{covid_chile}
\alias{covid_chile}
\title{Casos nuevos diarios Covid-19 en Chile.}
\usage{
covid_chile()
}
\arguments{
\item{NULL}{Esta funcion no usa parametros.}
}
\value{
Retorna los contagiados de los ultimos 7 dias y un grafico.
}
\description{
Casos nuevos diarios Covid-19 en Chile.
}
\examples{
covid_chile()
}
